refactor(api): move StageAcquisition model to evaluation scope
